pyproject.toml
[project]
name = "langgraph-basics"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "haongnd" }
]
requires-python = ">=3.11"
dependencies = [
    "langchain-community>=0.3.21",
    "langchain-openai>=0.3.12",
    "langgraph>=0.3.28",
    "openai>=1.72.0",
    "python-dotenv>=1.1.0",
]

[project.scripts]
langgraph-basics = "langgraph_basics:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[dependency-groups]
dev = [
    "ipykernel>=6.29.5",
    "ipython>=9.1.0",
]
